LOG-2201: fix loglevel matchers
