+++ internal/generator/fluentd/ingress.go
@@ -285,19 +285,19 @@ const ViaQDataModel string = `
   process_kubernetes_events 'false'
   <level>
     name warn
-    match 'Warning|WARN|W[0-9]+|level=warn|Value:warn|"level":"warn"'
+    match 'Warning|WARN|^W[0-9]+|level=warn|Value:warn|"level":"warn"'
   </level>
   <level>
     name info
-    match 'Info|INFO|I[0-9]+|level=info|Value:info|"level":"info"'
+    match 'Info|INFO|^I[0-9]+|level=info|Value:info|"level":"info"'
   </level>
   <level>
     name error
-    match 'Error|ERROR|E[0-9]+|level=error|Value:error|"level":"error"'
+    match 'Error|ERROR|^E[0-9]+|level=error|Value:error|"level":"error"'
   </level>
   <level>
     name debug
-    match 'Debug|DEBUG|D[0-9]+|level=debug|Value:debug|"level":"debug"'
+    match 'Debug|DEBUG|^D[0-9]+|level=debug|Value:debug|"level":"debug"'
   </level>
   <formatter>
     tag "system.var.log**"
